Relocated reflection methods
- Relocated methods to subclass named Reflections at ElementAcquisition
