+++ src/main/java/pickleib/element/ElementAcquisition.java
@@ -5,13 +5,16 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import org.openqa.selenium.By;
+import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.pagefactory.ByAll;
 import pickleib.enums.PrimarySelectorType;
 import pickleib.enums.SelectorType;
+import pickleib.exceptions.PickleibException;
 import pickleib.utilities.WebUtilities;
 import records.Bundle;
 import records.Pair;
+import utils.ReflectionUtilities;
 
 import java.util.*;
 
@@ -21,6 +24,8 @@ import static utils.StringUtilities.Color.*;
 public class ElementAcquisition {
 
     public static class PageObjectModel extends WebUtilities {
+        Reflections reflections = new Reflections();
+
         /**
          *
          * Acquire element {element name} from {page name}
@@ -37,7 +42,7 @@ public class ElementAcquisition {
             );
             pageName = strUtils.firstLetterDeCapped(pageName);
             elementName = strUtils.contextCheck(elementName);
-            return getElementFromPage(elementName, pageName, objectRepository);
+            return reflections.getElementFromPage(elementName, pageName, objectRepository);
         }
 
         /**
@@ -58,7 +63,7 @@ public class ElementAcquisition {
             pageName = strUtils.firstLetterDeCapped(pageName);
             componentFieldName = strUtils.firstLetterDeCapped(componentFieldName);
             elementName = strUtils.contextCheck(elementName);
-            return getElementFromComponent(elementName, componentFieldName, pageName, objectRepository);
+            return reflections.getElementFromComponent(elementName, componentFieldName, pageName, objectRepository);
         }
 
         /**
@@ -87,7 +92,7 @@ public class ElementAcquisition {
             pageName = strUtils.firstLetterDeCapped(pageName);
             listName = strUtils.firstLetterDeCapped(listName);
             elementName = strUtils.contextCheck(elementName);
-            List<WebElement> elements = getElementsFromPage(
+            List<WebElement> elements = reflections.getElementsFromPage(
                     listName,
                     pageName,
                     objectRepository
@@ -126,7 +131,7 @@ public class ElementAcquisition {
             pageName = strUtils.firstLetterDeCapped(pageName);
             listFieldName = strUtils.firstLetterDeCapped(listFieldName);
             elementName = strUtils.contextCheck(elementName);
-            List<WebElement> elements = getElementsFromComponent(
+            List<WebElement> elements = reflections.getElementsFromComponent(
                     listFieldName,
                     componentFieldName,
                     pageName,
@@ -164,9 +169,9 @@ public class ElementAcquisition {
             pageName = strUtils.firstLetterDeCapped(pageName);
             componentListName = strUtils.firstLetterDeCapped(componentListName);
             elementName = strUtils.contextCheck(elementName);
-            List<WebComponent> componentList = getComponentsFromPage(componentListName, pageName, objectRepository);
+            List<WebComponent> componentList = reflections.getComponentsFromPage(componentListName, pageName, objectRepository);
             WebComponent component = acquireNamedComponentAmongst(componentList, componentName);
-            return getElementFromComponent(elementName, component);
+            return reflections.getElementFromComponent(elementName, component);
         }
 
         /**
@@ -194,7 +199,7 @@ public class ElementAcquisition {
                     highlighted(BLUE, pageName)
             );
             WebComponent component = exactNamedListedComponent(elementFieldName, elementText, componentListName, pageName, objectRepository);
-            return getElementFromComponent(elementFieldName, component);
+            return reflections.getElementFromComponent(elementFieldName, component);
         }
 
         /**
@@ -224,8 +229,8 @@ public class ElementAcquisition {
             pageName = strUtils.firstLetterDeCapped(pageName);
             componentListName = strUtils.firstLetterDeCapped(componentListName);
             elementFieldName = strUtils.contextCheck(elementFieldName);
-            List<WebComponent> components = getComponentsFromPage(componentListName, pageName, objectRepository);
-            return acquireExactNamedComponentAmongst(components, elementText, elementFieldName);
+            List<WebComponent> components = reflections.getComponentsFromPage(componentListName, pageName, objectRepository);
+            return reflections.acquireExactNamedComponentAmongst(components, elementText, elementFieldName);
         }
 
         /**
@@ -258,9 +263,9 @@ public class ElementAcquisition {
             componentName = strUtils.contextCheck(componentName);
             pageName = strUtils.firstLetterDeCapped(pageName);
             componentListName = strUtils.firstLetterDeCapped(componentListName);
-            List<WebComponent> components = getComponentsFromPage(componentListName, pageName, objectRepository);
+            List<WebComponent> components = reflections.getComponentsFromPage(componentListName, pageName, objectRepository);
             WebComponent component = acquireNamedComponentAmongst(components, componentName);
-            List<WebElement> elements = getElementsFromComponent(elementListName, component);
+            List<WebElement> elements = reflections.getElementsFromComponent(elementListName, component);
             return acquireNamedElementAmongst(elements, elementName);
         }
 
@@ -290,7 +295,7 @@ public class ElementAcquisition {
             );
             attributeName = strUtils.contextCheck(attributeName);
             pageName = strUtils.firstLetterDeCapped(pageName);
-            List<WebElement> elements = getElementsFromPage(
+            List<WebElement> elements = reflections.getElementsFromPage(
                     listName,
                     strUtils.firstLetterDeCapped(pageName),
                     objectRepository
@@ -327,7 +332,7 @@ public class ElementAcquisition {
             attributeName = strUtils.contextCheck(attributeName);
             pageName = strUtils.firstLetterDeCapped(pageName);
             componentName = strUtils.firstLetterDeCapped(componentName);
-            List<WebElement> elements = getElementsFromComponent(
+            List<WebElement> elements = reflections.getElementsFromComponent(
                     listName,
                     componentName,
                     pageName,
@@ -351,7 +356,7 @@ public class ElementAcquisition {
                 String inputName = form.get("Input Element");
                 String input = strUtils.contextCheck(form.get("Input"));
                 Bundle<WebElement, String, String> bundle = new Bundle<>(
-                        getElementFromPage(inputName, pageName, objectRepository),
+                        reflections.getElementFromPage(inputName, pageName, objectRepository),
                         input,
                         inputName
                 );
@@ -376,7 +381,7 @@ public class ElementAcquisition {
                 String input = strUtils.contextCheck(form.get("Input"));
                 componentName = strUtils.firstLetterDeCapped(componentName);
                 Bundle<WebElement, String, String> bundle = new Bundle<>(
-                        getElementFromComponent(inputName, componentName, pageName, objectRepository),
+                        reflections.getElementFromComponent(inputName, componentName, pageName, objectRepository),
                         input,
                         inputName
                 );
@@ -502,9 +507,9 @@ public class ElementAcquisition {
             String selectorElementName = strUtils.contextCheck(specifications.get("Selector Element"));
             String targetElementName = strUtils.contextCheck(specifications.get("Target Element"));
             pageName = strUtils.firstLetterDeCapped(pageName);
-            List<WebComponent> components = getComponentsFromPage(componentListName, pageName, objectRepository);
-            WebComponent component = acquireExactNamedComponentAmongst(components, selectorElementText, selectorElementName);
-            return new Bundle<>(targetElementName, getElementFromComponent(targetElementName, component), specifications);
+            List<WebComponent> components = reflections.getComponentsFromPage(componentListName, pageName, objectRepository);
+            WebComponent component = reflections.acquireExactNamedComponentAmongst(components, selectorElementText, selectorElementName);
+            return new Bundle<>(targetElementName, reflections.getElementFromComponent(targetElementName, component), specifications);
         }
 
         /**
@@ -707,5 +712,343 @@ public class ElementAcquisition {
 
             return null;
         }
+    }
+
+    public static class Reflections extends WebUtilities {
+        protected ReflectionUtilities reflectionUtils = new ReflectionUtilities();
+
+        /**
+         * Acquires an element from a given page
+         *
+         * @param elementFieldName element field name
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns the element
+         */
+        protected WebElement getElementFromPage(String elementFieldName, String pageName, Object objectRepository){
+            Map<String, Object> pageFields;
+            Object pageObject = reflectionUtils.getFields(objectRepository).get(pageName);
+            if (pageObject != null) pageFields = reflectionUtils.getFields(pageObject);
+            else throw new PickleibException("ObjectRepository does not contain an instance of " + pageName + " object!");
+            return (WebElement) pageFields.get(elementFieldName);
+        }
+
+        /**
+         * Acquires a list of elements from a given page
+         *
+         * @param elementFieldName element field name
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns the list of elements
+         */
+        @SuppressWarnings("unchecked")
+        protected List<WebElement> getElementsFromPage(String elementFieldName, String pageName, Object objectRepository){
+            Map<String, Object> pageFields;
+            Object pageObject = reflectionUtils.getFields(objectRepository).get(pageName);
+            if (pageObject != null) pageFields = reflectionUtils.getFields(pageObject);
+            else throw new PickleibException("ObjectRepository does not contain an instance of " + pageName + " object!");
+            return (List<WebElement>) pageFields.get(elementFieldName);
+        }
+
+        /**
+         * Acquires an element from a given component
+         *
+         * @param elementFieldName element field name
+         * @param selectionName element text
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns the element
+         */
+        protected WebElement getElementAmongstComponentsFromPage(
+                String elementFieldName,
+                String selectionName,
+                String componentListName,
+                String pageName,
+                Object objectRepository){
+            List<WebComponent> componentList = getComponentsFromPage(componentListName, pageName, objectRepository);
+            WebComponent component = acquireNamedComponentAmongst(componentList, selectionName);
+            Map<String, Object> componentFields = reflectionUtils.getFields(component);
+            return (WebElement) componentFields.get(elementFieldName);
+        }
+
+        /**
+         * Acquires a list of elements from a given component
+         *
+         * @param elementFieldName element field name
+         * @param selectionName element text
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns the list of elements
+         */
+        @SuppressWarnings("unchecked")
+        protected List<WebElement> getElementsAmongstComponentsFromPage(
+                String elementFieldName,
+                String selectionName,
+                String componentListName,
+                String pageName,
+                Object objectRepository){
+            List<WebComponent> componentList = getComponentsFromPage(componentListName, pageName, objectRepository);
+            WebComponent component = acquireNamedComponentAmongst(componentList, selectionName);
+            Map<String, Object> componentFields = reflectionUtils.getFields(component);
+            return (List<WebElement>) componentFields.get(elementFieldName);
+        }
+
+        /**
+         * Acquires an element from a component amongst a list of components
+         *
+         * @param elementFieldName element field name
+         * @param selectionName element text
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns the element
+         */
+        protected WebElement getElementAmongstNamedComponentsFromPage(
+                String elementFieldName,
+                String selectionName,
+                String componentListName,
+                String pageName,
+                Object objectRepository){
+            List<WebComponent> componentList = getComponentsFromPage(componentListName, pageName, objectRepository);
+            WebComponent component = acquireNamedComponentAmongst(componentList, selectionName);
+            Map<String, Object> componentFields = reflectionUtils.getFields(component);
+            return (WebElement) componentFields.get(elementFieldName);
+        }
+
+        /**
+         * Acquires a list of elements from a component amongst a list of components
+         *
+         * @param listFieldName list field name
+         * @param selectionName element text
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns the list of elements
+         */
+        @SuppressWarnings("unchecked")
+        protected List<WebElement> getElementsAmongstNamedComponentsFromPage(
+                String listFieldName,
+                String selectionName,
+                String componentListName,
+                String pageName,
+                Object objectRepository){
+            List<WebComponent> componentList = getComponentsFromPage(componentListName, pageName, objectRepository);
+            WebComponent component = acquireNamedComponentAmongst(componentList, selectionName);
+            Map<String, Object> componentFields = reflectionUtils.getFields(component);
+            return (List<WebElement>) componentFields.get(listFieldName);
+        }
+
+        /**
+         * Acquires an element from a component amongst a list of components
+         *
+         * @param elementFieldName element field name
+         * @param elementIdentifier element text
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns the element if exact match found
+         */
+        @Deprecated(since = "1.6.2")
+        protected WebElement getElementAmongstExactComponentsFromPage(
+                String elementFieldName,
+                String elementIdentifier,
+                String componentListName,
+                String pageName,
+                Object objectRepository){
+            WebComponent component = acquireExactNamedComponentAmongst(elementIdentifier, elementFieldName, componentListName, pageName, objectRepository);
+            Map<String, Object> componentFields = reflectionUtils.getFields(component);
+            return (WebElement) componentFields.get(elementFieldName);
+        }
+
+        /**
+         * Acquires a map of fields from a given component
+         *
+         * @param componentName component name
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns map of fields
+         */
+        protected Map<String, Object> getComponentFieldsFromPage(String componentName, String pageName, Object objectRepository){
+            Map<String, Object> componentFields;
+            Object pageObject = reflectionUtils.getFields(objectRepository).get(pageName);
+            if (pageObject != null) componentFields = reflectionUtils.getFields(pageObject);
+            else throw new PickleibException("ObjectRepository does not contain an instance of " + pageName + " object!");
+            return reflectionUtils.getFields(componentFields.get(componentName));
+        }
+
+        /**
+         * Acquires a list of element from a given page
+         *
+         * @param componentListName component list name
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns the list of components
+         */
+        @SuppressWarnings("unchecked")
+        protected List<WebComponent> getComponentsFromPage(String componentListName, String pageName, Object objectRepository){
+            Map<String, Object> pageFields;
+            Map<String, Object> componentFields;
+            Object pageObject = reflectionUtils.getFields(objectRepository).get(pageName);
+            if (pageObject != null) pageFields = reflectionUtils.getFields(pageObject);
+            else throw new PickleibException("ObjectRepository does not contain an instance of " + pageName + " object!");
+            return (List<WebComponent>) pageFields.get(componentListName);
+        }
+
+        /**
+         * Acquire a map of fields from a given component
+         *
+         * @param componentName component name
+         * @return returns the map of fields
+         */
+        protected Map<String, Object> getComponentFields(Object componentName){
+            return  reflectionUtils.getFields(componentName);
+        }
+
+        /**
+         * Acquire listed component by the text of its given child element
+         *
+         * @param items list of components
+         * @param attributeName component element attribute name
+         * @param attributeValue attribute value
+         * @param elementFieldName component elements field name
+         * @return returns the matching component
+         * @param <T> component type
+         */
+        protected  <T> T acquireComponentByElementAttributeAmongst(
+                List<T> items,
+                String attributeName,
+                String attributeValue,
+                String elementFieldName
+        ){
+            log.info("Acquiring component by attribute " + strUtils.highlighted(BLUE, attributeName + " -> " + attributeValue));
+            boolean timeout = false;
+            long initialTime = System.currentTimeMillis();
+            while (!timeout){
+                for (T component : items) {
+                    Map<String, Object> componentFields = reflectionUtils.getFields(component);
+                    WebElement element = (WebElement) componentFields.get(elementFieldName);
+                    String attribute = element.getAttribute(attributeName);
+                    if (attribute.equals(attributeValue)) return component;
+                }
+                if (System.currentTimeMillis() - initialTime > elementTimeout) timeout = true;
+            }
+            throw new NoSuchElementException("No component with " + attributeName + " : " + attributeValue + " could be found!");
+        }
+
+        /**
+         * Acquire listed component by the text of its given child element
+         *
+         * @param items list of components
+         * @param elementText text of the component element
+         * @param targetElementFieldName component elements field name
+         * @return returns the matching component
+         * @param <Component> component type
+         */
+        protected  <Component extends WebComponent> Component acquireExactNamedComponentAmongst(
+                List<Component> items,
+                String elementText,
+                String targetElementFieldName
+        ){
+            log.info("Acquiring component called " + strUtils.highlighted(BLUE, elementText));
+            boolean timeout = false;
+            long initialTime = System.currentTimeMillis();
+            while (!timeout){
+                for (Component component : items) {
+                    Map<String, Object> componentFields = reflectionUtils.getFields(component);
+                    WebElement element = (WebElement) componentFields.get(targetElementFieldName);
+                    String text = element.getText();
+                    String name = element.getAccessibleName();
+                    if (text.equalsIgnoreCase(elementText) || name.equalsIgnoreCase(elementText)) return component;
+                }
+                if (System.currentTimeMillis() - initialTime > elementTimeout) timeout = true;
+            }
+            throw new NoSuchElementException("No component with text/name '" + elementText + "' could be found!");
+        }
+
+        /**
+         *
+         * @deprecated replaced by acquireExactNamedComponentAmongst(components, elementText, elementFieldName)
+         */
+        @Deprecated(since = "1.6.2")
+        protected WebComponent acquireExactNamedComponentAmongst(
+                String elementText,
+                String elementFieldName,
+                String componentListName,
+                String pageName,
+                Object objectRepository){
+            log.info("Acquiring component called " + strUtils.highlighted(BLUE, elementText));
+            boolean timeout = false;
+            long initialTime = System.currentTimeMillis();
+            while (!timeout){
+                for (WebComponent component : getComponentsFromPage(componentListName, pageName, objectRepository)) {
+                    Map<String, Object> componentFields = reflectionUtils.getFields(component);
+                    WebElement element = (WebElement) componentFields.get(elementFieldName);
+                    String text = element.getText();
+                    String name = element.getAccessibleName();
+                    if (text.equalsIgnoreCase(elementText) || name.equalsIgnoreCase(elementText)) return component;
+                }
+                if (System.currentTimeMillis() - initialTime > elementTimeout) timeout = true;
+            }
+            throw new NoSuchElementException("No component with text/name '" + elementText + "' could be found!");
+        }
+
+        /**
+         * Acquires web element from a page object by using Java reflections
+         *
+         * @param fieldName field name of the element, in the page object
+         * @param inputClass instance of the page object that the WebElement resides in
+         * @return corresponding WebElement from the given page object
+         */
+        protected  <T> WebElement getElement(String fieldName, Class<T> inputClass){
+            return (WebElement) reflectionUtils.getFieldValue(fieldName, inputClass);
+        }
+
+        /**
+         * Acquires an element from a given component name
+         *
+         * @param elementFieldName element field name
+         * @param componentName target component
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns the element
+         */
+        protected WebElement getElementFromComponent(String elementFieldName, String componentName, String pageName, Object objectRepository){
+            return (WebElement) getComponentFieldsFromPage(componentName, pageName, objectRepository).get(elementFieldName);
+        }
+
+        /**
+         * Acquires an element from a given component
+         *
+         * @param elementFieldName element field name
+         * @param component target component
+         * @return returns the element
+         */
+        protected WebElement getElementFromComponent(String elementFieldName, WebComponent component){
+            return (WebElement) getComponentFields(component).get(elementFieldName);
+        }
+
+        /**
+         * Acquires a list elements from a given component name
+         *
+         * @param listFieldName element field
+         * @param componentName target component name
+         * @param pageName name of the page instance
+         * @param objectRepository instance of an object that contains instances of every page
+         * @return returns the list of elements
+         */
+        @SuppressWarnings("unchecked")
+        protected List<WebElement> getElementsFromComponent(String listFieldName, String componentName, String pageName, Object objectRepository){
+            return (List<WebElement>) getComponentFieldsFromPage(componentName, pageName, objectRepository).get(listFieldName);
+        }
+
+        /**
+         * Acquires a list elements from a given component
+         *
+         * @param elementListFieldName elements list field
+         * @param component target component
+         * @return returns the list of elements
+         */
+        @SuppressWarnings("unchecked")
+        protected List<WebElement> getElementsFromComponent(String elementListFieldName, Object component){
+            return (List<WebElement>) getComponentFields(component).get(elementListFieldName);
+        }
+
     }
 }
